feat(sidebar): resizable, retractable sidebar with grouped items
- Resizable sidebar
- Retractable to a smaller sidebar
- Grouped sidebar items
    - Management group for users, Services, nodes, host
- kebab convention for folder and file system
