+++ app/dashboard/src/components/sidebar/logo.tsx
@@ -1,0 +1,7 @@
+import { Image } from '@chakra-ui/react'
+
+export const Logo = () => {
+    return (
+        <Image src="favicon/logo.svg" alt="marzneshin" boxSize="50px" w="full" />
+    )
+}
